Move CICP module to metadata
