+++ src/color.rs
@@ -1,5 +1,3 @@
-pub(crate) mod cicp;
-
 use std::ops::{Index, IndexMut};
 
 use num_traits::{NumCast, ToPrimitive, Zero};
