+++ src/lib.rs
@@ -138,11 +138,6 @@ pub use crate::color::{ColorType, ExtendedColorType};
 
 pub use crate::color::{Luma, LumaA, Rgb, Rgba};
 
-pub use crate::color::cicp::{
-    Cicp, CicpColorPrimaries, CicpMatrixCoefficients, CicpTransferFunction, CicpTransform,
-    CicpVideoFullRangeFlag,
-};
-
 pub use crate::error::{ImageError, ImageResult};
 
 pub use crate::images::generic_image::{GenericImage, GenericImageView, Pixels};
